working on player info card
